% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plothelp.R
\name{E3Axes}
\alias{E3Axes}
\title{Draw axes and labels on x/y, using Rcssplot}
\usage{
E3Axes(xlim, ylim, xlab = "", ylab = "", xat = NULL, RC = "default",
  RCC = c(), where = c("x", "y"))
}
\arguments{
\item{xlim}{- two element vector. Range of x axis}

\item{ylim}{- two element vector. Range of y axis}

\item{xlab}{- character string. Text to write below x-axis}

\item{ylab}{- character string. Text to write below y-axis}

\item{xat}{- numeric vector. Position of labels on x-axis.}

\item{RC}{- Rcss object. Style for formatting the axes using Rcssplot}

\item{RCC}{- character vector. Classes for tuning Rcssplot formating}

\item{where}{- vector containg "x" and/or "y". Determines where axes will be drawn}
}
\description{
Draw axes and labels on x/y, using Rcssplot
}

