% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-rescale.R
\name{E3CalcAcrossGroupIntervalRescaleFactors}
\alias{E3CalcAcrossGroupIntervalRescaleFactors}
\title{Estimate an interval rescale factor for each feature/gene based on across-group
variability.}
\usage{
E3CalcAcrossGroupIntervalRescaleFactors(sampledata, id.col = "Gene",
  bg.samples = NULL, upper.quantile = 0.85, min.factor = 1)
}
\arguments{
\item{sampledata}{- a list with tables "expression", "expression.low", "expression.high"}

\item{id.col}{- character strings. Name of column in sampledata that identifies
the gene/transcript species}

\item{bg.samples}{- list of sample names. Each vector should represent a group of related
samples (replicates). The function will compute variability between across these groups.}

\item{upper.quantile}{- numeric. The quantile used to equalize the input and output intervals.}

\item{min.factor}{- numeric. Minimum rescaling factor, i.e. when procedure gives
a suggestion to rescale by a factor lower than this, the output will in fact be min.factor.}
}
\description{
This goal is to find genes that are prone to batch effects, i.e. that differ between
batches more than their intervals would suggest.
}

