% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-rescale.R
\name{E3CalcWithinGroupIntervalRescaleFactors}
\alias{E3CalcWithinGroupIntervalRescaleFactors}
\title{Estimates an interval rescale factor for each feature/gene based on within-group variability.}
\usage{
E3CalcWithinGroupIntervalRescaleFactors(sampledata, id.col = "Gene",
  bg.samples = NULL, upper.quantile = 0.85, min.factor = 1)
}
\arguments{
\item{sampledata}{- a list with tables $expression, $expression.low, $expression.high}

\item{id.col}{- character strings. Name of column in sampledata that identifies
the gene/transcript species}

\item{bg.samples}{- list of sample names. Each vector should represent a group of related
samples (replicates). The function will compute variability within each group.}

\item{upper.quantile}{- numeric. The quantile used to equalize the input and output intervals.}

\item{min.factor}{- numeric. Minimum rescaling factor, i.e. when procedure gives
a suggestion to rescale by a factor lower than this, the output will in fact be min.factor.}
}
\description{
The goal is to find genes that are more variable in practice than the
uncertainty interval model may suggest.
}
\details{
computes an empirical interval from values of a background sample set
computes a representative interval from values of the background sample set
computes the ratio of the empirical to representative - this is the output
}

