% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3CompareSignatures}
\alias{E3CompareSignatures}
\title{Collect metrics (Jaccard, Overlap, Fisher) describing to what extent observed
gene signatures correspond to expected signatures.}
\usage{
E3CompareSignatures(sampleGroups, group.signature = NULL,
  signature.ref = NULL, signature.expected = NULL, numgenes = 5000)
}
\arguments{
\item{sampleGroups}{- named list of character vectors. Each vector should represent
a set of samples making up a group.}

\item{group.signature}{- named list of character vectors. Each vector should represent
the gene set (signature) associated with a group}

\item{signature.ref}{- named character vector. Each element should link a sample name
to a reference group, e.g. KO_ACTA_R1 would be linked to group KO_None}

\item{signature.expected}{- named character vector. Each element should a sample
to an expected signature group, e.g. KO_ACTA_R1 would be linked to group WT_ACTA}

\item{numgenes}{- number of genes to use as background in calculation of Fisher metric.}
}
\description{
Metrics include overlap between observed and expected signatures, JI between observed
and expected signatures, and Fisher p value for selection of observed and expected
signatures from a background set.
}
\details{
Warning: the Fisher p value is computed very crudely using a background set of fixed size.
This is in principle flawed, but the whole idea of using a Fisher test in such a scenario
is flawed because not all genes are equally powered to be part of signatures. Bottom line
is that the Fisher p values should be interpreted with care, they may be off even by
two/three orders of magnitude.
}

