% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3CompareVectors}
\alias{E3CompareVectors}
\title{take two vectors A, B
returns a list with elements in A only, B only, and both A and B.}
\usage{
E3CompareVectors(A = c(), B = c())
}
\arguments{
\item{A}{- vector. First vectors to compare}

\item{B}{- vector. Second vector to compare}
}
\description{
take two vectors A, B
returns a list with elements in A only, B only, and both A and B.
}

