% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-help.R
\name{E3DEscore}
\alias{E3DEscore}
\title{Compute a score summarizing a fold-change and a z-value into one number [-1, 1]}
\usage{
E3DEscore(fc, z, fc.thresh = c(1.25, 1.75), z.thresh = c(1.25, 1.75))
}
\arguments{
\item{fc}{numeric. Values of fold change (e.g. 2 for 2-fold up-regualtion,
0.25 for 4-fold down-regulation).}

\item{z}{numeric. Values of z-score}

\item{fc.thresh}{numeric vector of length two. Should hold lower and upper thresholds for
interesting fold-changes. Values of fc below lower threshold will receive score 0. Values above
upper threshold will receive score 1. Intermediate values follow linear interpolation.}

\item{z.thresh}{numeric vector of length two. Analogous to fc.thresh, but here thresholds
applied to z score.}
}
\description{
This is a suggested scoring function that collapses two informative numbers, a fold
change conveying effect size and a z-score conveying significance, into a single number
in a fixed range [-1, 1]
}

