% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-help.R
\name{E3dist.euclidean}
\alias{E3dist.euclidean}
\title{Compute euclidean distances between all pairs of samples}
\usage{
E3dist.euclidean(dd, samplenames)
}
\arguments{
\item{dd}{- numeric matrix. Columns should contain sample names (next argument)}

\item{samplenames}{- character vector with column names in dd.}
}
\description{
Compute euclidean distances between all pairs of samples
}

