% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-help.R
\name{E3dist.spearman}
\alias{E3dist.spearman}
\title{Compute spearman pseudo-distances between all pairs of samples. Spearman
pseudo distance between two samples is defined as 1-spearman.rho}
\usage{
E3dist.spearman(dd, samplenames)
}
\arguments{
\item{dd}{- numeric matrix. Columns should contain sample names (next argument)}

\item{samplenames}{- character vector with column names}
}
\description{
Compute spearman pseudo-distances between all pairs of samples. Spearman
pseudo distance between two samples is defined as 1-spearman.rho
}

