% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plothelp.R
\name{E3DrawMarkerLegend}
\alias{E3DrawMarkerLegend}
\title{Draw a vertical legend with markers (circles, boxes, and labels)}
\usage{
E3DrawMarkerLegend(legend = NULL, xlim = c(0, 1), ylim = c(0, 1),
  main = "", RC = "default", RCC = c())
}
\arguments{
\item{legend}{- data frame with columns cex, pch, col, bg, and label. Order and contents
determines what markers are drawn}

\item{xlim}{- numeric vector of two elements. Range of x axis (used to place the legend)}

\item{ylim}{- numeric vector of two elements. Range of y axis (used to place the legend)}

\item{main}{- text to write above the legend}

\item{RC}{- Rcss object. Style to use for plotting, uses package Rcssplot}

\item{RCC}{- character vetor. Classes to tune Rcssplot formatting.}
}
\description{
This function uses Rcss object "markerlegend" to determine legend position.
}

