% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-makesigs.R
\name{E3GetDEScores}
\alias{E3GetDEScores}
\title{Compute DE scores for a single sample.}
\usage{
E3GetDEScores(sampledata, sampleA, refsamples, scorefun = E3DEscore,
  rankrescale = TRUE, shiftval = 0.1)
}
\arguments{
\item{sampledata}{- list with expression, expression.low, expression.high tables}

\item{sampleA}{- character string. The sample for which the signature is made}

\item{refsamples}{- character vector. Reference samples to which sampleA is compared.}

\item{scorefun}{- function. This function should accept two arguments (fc and z) and return
a numeric score.}

\item{rankrescale}{- logical. Set TRUE to penalize scores in an ad-hoc way for
genes that have a high score but lower fc or lower z than many other genes.}

\item{shiftval}{- numeric. Used when computing FC to avoid division by zero.
return - a vector with gene names and then a score [0,1]
         Score 1 means the gene is DE wrt all samples in groupB}
}
\description{
Function returns a named vector with scores [-1,1]. Names are genes. Values are
scores that indicate whether a gene is DE based on FC and z criteria.
}

