% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-help.R
\name{E3GetExpressedGenes}
\alias{E3GetExpressedGenes}
\title{Determine genes that are expressed in a dataset. Expressed genes are those that
have expression in a range (min.expression, max.expression)}
\usage{
E3GetExpressedGenes(dlist, id.col = "Gene", min.expression = 1,
  max.expression = Inf, min.count = 1)
}
\arguments{
\item{dlist}{- a list of dataframes, one of which must be $expression. Data frames
must have one identifier column and the remaining columns will be treated as samples.}

\item{id.col}{- character string. Name of column that contain gene identifier}

\item{min.expression}{- numeric. Minimal expression level}

\item{max.expression}{- numeric. Maximal expression level}

\item{min.count}{- integer. Number of samples that must lie in range
[min.expression, max.expression]. Use this to require a gene to be expressed
in several samples in a cohort.}
}
\description{
Determine genes that are expressed in a dataset. Expressed genes are those that
have expression in a range (min.expression, max.expression)
}

