% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3GetFisherP}
\alias{E3GetFisherP}
\title{Compute fisher p-value: selection of two vectors from a background set}
\usage{
E3GetFisherP(x, y, background)
}
\arguments{
\item{x}{- vector}

\item{y}{- vector}

\item{background}{- large vector from which x and y were picked}
}
\description{
Compute fisher p-value: selection of two vectors from a background set
}

