% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-segmentation.R
\name{E3GetGenePosition}
\alias{E3GetGenePosition}
\title{Obtain genomic positions of genes from a refGene-style annotation table.}
\usage{
E3GetGenePosition(annotab, chrnames = NULL)
}
\arguments{
\item{annotab}{- data frame. A refGene style table with transcript definitions}

\item{chrnames}{- character vector. Use this to look at transcripts on just a few
chromosomes, e.g. to avoid genes on decoy contigs.}
}
\description{
Obtain genomic positions of genes from a refGene-style annotation table.
}

