% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3GetJaccard}
\alias{E3GetJaccard}
\title{Compute Jaccard index of two vectors (size of intersection divided by sum of union)}
\usage{
E3GetJaccard(x, y)
}
\arguments{
\item{x}{- vector}

\item{y}{-vector}
}
\description{
Compute Jaccard index of two vectors (size of intersection divided by sum of union)
}

