% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3GetJaccardDistance}
\alias{E3GetJaccardDistance}
\title{Compute a distance matrix between gene sets based on the Jaccard index.}
\usage{
E3GetJaccardDistance(sigs)
}
\arguments{
\item{sigs}{- named list of character vectors. Each vector should represent a
gene set. The function will compute pairwise Jaccard Indexes between these vectors.}
}
\description{
The function will compute pairwise Jaccard Index (JI) between gene sets. The distance
between the gene sets will be 1-JI.
}

