% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-FC.R
\name{E3GetObservedExpectedGroupFC}
\alias{E3GetObservedExpectedGroupFC}
\title{Produce a score based on whether a genes is consistent with expected behavior (fold changes)}
\usage{
E3GetObservedExpectedGroupFC(sampleexp, sampleGroups, signature.ref,
  signature.expected, fun = mean, add.dark = 1, id.col = "Gene")
}
\arguments{
\item{sampleexp}{- data frame with columns (Gene, SAMPLENAMES)}

\item{sampleGroups}{- named list of character vectors (sample names). Names indicate
sample grouping, e.g. group WT_None3 containing samples c("WT_None3_R1", "WT_None2_R2")}

\item{signature.ref}{- named character vector. This should link samplenames to
a group with reference samples, e.g. sample KO_ACTA_R1 will link to a reference group
KO_None.}

\item{signature.expected}{- names character vector. This should link samplenames to
expected signature groups, e.g. sample KO_ACTA_R1 will be expected to be similar to
group WT_ACTA.}

\item{fun}{- function. Used to summarize data from multiple samples into one group value.
The mean summary function is usually appropriate and fast.}

\item{add.dark}{- numeric. a dark count for the expression values. This is useful to avoid
division by zeros.}

\item{id.col}{- name of column in sampleexp with gene identifier}
}
\description{
Function returns a list with two tables. Values in one table are expected fold
changes. Values in the other tables are observed fold changes. See descriptions
of arguments to see how these are computed. Each table will have gene names in rownames
and group names in colnames.
}

