% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3GetOverlap}
\alias{E3GetOverlap}
\title{Compute overlap (number of shared elements) between two vectors x, y}
\usage{
E3GetOverlap(x, y)
}
\arguments{
\item{x}{- vector}

\item{y}{- vector}
}
\description{
Compute overlap (number of shared elements) between two vectors x, y
}

