% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3GetRelativeOverlap}
\alias{E3GetRelativeOverlap}
\title{Compute relative overlap (number of shared elements divided by size of first)
between two vectors x, y}
\usage{
E3GetRelativeOverlap(x, y)
}
\arguments{
\item{x}{- vector}

\item{y}{- vector}
}
\description{
Compute relative overlap (number of shared elements divided by size of first)
between two vectors x, y
}

