% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-load.R
\name{E3LoadExp3pGeneData}
\alias{E3LoadExp3pGeneData}
\title{Load expression data for samples from Exp3p-like data files.}
\usage{
E3LoadExp3pGeneData(samplenames, platenames, path.template,
  column.template = "SAMPLE.EPM", column.low.template = "SAMPLE.EPM.low",
  column.high.template = "SAMPLE.EPM.high", id.in = "Region.id",
  id.out = "Gene", rescale.uncertainty = 1, verbose = TRUE)
}
\arguments{
\item{samplenames}{- character vector with sample codes e.g. KO_None. The sample codes
must correspond to parts of the data path.}

\item{platenames}{- character vector with additional codes for each sample, e.g. a plate id.
These codes must correspond to parts of the data path.}

\item{path.template}{- character string. This is a generic path that can contain patterns
SAMPLE and PLATE. These substrings will be replaced by values from the previous two vectors.}

\item{column.template}{- character string marking the column in the data table containing
values to be interpreted as expression values. The pattern SAMPLE will be replaced by
values from the samplenames vector.}

\item{column.low.template}{- character string. Similar to column.template, but referring
to columns with a low estimate for expression values}

\item{column.high.template}{- character string. Similar to column.template, but referring
to columns with a high estimate for expression values}

\item{id.in}{- character string. Name of column in input data files containing the
gene or region identifier.}

\item{id.out}{- character string. Name of column in the output data frame containing
gene identifier}

\item{rescale.uncertainty}{- numeric. When different from one, the low/high estimates
will be modified by this factor. Use this value to read more conservative
expression intervals.}

\item{verbose}{- logical. Print out progress messages.}
}
\description{
Load expression data for samples from Exp3p-like data files.
}

