% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-makesigs.R
\name{E3LoadGMT}
\alias{E3LoadGMT}
\title{Read GMT signatures from a file and produce a list of vectors with gene names}
\usage{
E3LoadGMT(f)
}
\arguments{
\item{f}{- filename}
}
\description{
Read GMT signatures from a file and produce a list of vectors with gene names
}

