% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-segmentation.R
\name{E3MakeFCSegmentation}
\alias{E3MakeFCSegmentation}
\title{Perform genomic segmentation.}
\usage{
E3MakeFCSegmentation(fcdata, gene.position = NULL, genenames = NULL,
  chrnames = NULL, min.width = 5, verbose = TRUE)
}
\arguments{
\item{fcdata}{- numeric matrix with sample names in colnames and gene names in rownames.}

\item{gene.position}{- data frame with columns "chr", "pos", and rownames
containing gene names.}

\item{genenames}{- names of}

\item{chrnames}{- character vector with names of chromosomes to include in segmentation.
Leave null, To use all the genes in the input matrix.}

\item{min.width}{- minimum number}

\item{verbose}{- logical. When TRUE, function prints dots indicating progress.}
}
\description{
This function uses CBS (package DNAcopy) to segment genomic data. This wrapper
allows to start with a data matrix containing values associated with gene names. A
separated object mapping genes to genomic positions is used to arrange the data
before segmentation.
}
\details{
Values in fcdata will be log2 transformed before being sent to segmentation. Functions
returns a list with the resulting segmentation.
}

