% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3MakeGLMs}
\alias{E3MakeGLMs}
\title{Create GLM models linking Overlap with an expected signature to technical covariates.}
\usage{
E3MakeGLMs(sigmetrics, group.covariates = NULL, glm.covariate = NULL,
  glm.readcounts = TRUE, depth.col = "Depth_M_reads")
}
\arguments{
\item{sigmetrics}{- a data frame with columns "Group", "Expected", "Overlap"}

\item{group.covariates}{-}

\item{glm.covariate}{- character. Name of column in group.covariates to use
in the glm model}

\item{glm.readcounts}{- logical. Set TRUE to include a term in the glm model
for read depth.}
}
\description{
This function is rather particular for ExpCube. It creates a series of GLM models,
one for each "Expected" signature type.
The function outputs a list of glm models and a new table analogous to sigmetrics.
}

