% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3MakeGroupCovariates}
\alias{E3MakeGroupCovariates}
\title{Collapse a per-sample table of covariates into a per-group table}
\usage{
E3MakeGroupCovariates(covariates = NULL, sampleGroups = NULL)
}
\arguments{
\item{covariates}{- data frame. Table with samplenames in rownames, then columns
representing categorial and/or numeric covariates}

\item{sampleGroups}{- named list of character vectors. Each vector represents the
samples that make up one group}
}
\description{
Function outputs a table with as many columns as covariates, as many rows as
groups defined in sampleGroups.
}

