% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-makesigs.R
\name{E3MakeGroupSignatureScores}
\alias{E3MakeGroupSignatureScores}
\title{Average scores over several replicates to obtain one score per group}
\usage{
E3MakeGroupSignatureScores(scores, sampleGroups)
}
\arguments{
\item{scores}{- matrix with scores. Column names should be sample names.}

\item{sampleGroups}{- named list with character vectors. Each vector will
define a set of samples to average into a group}
}
\description{
Average scores over several replicates to obtain one score per group
}

