% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-help.R
\name{E3MakeSampleGroups}
\alias{E3MakeSampleGroups}
\title{Create a list of vectors describing how samples are related into groups.}
\usage{
E3MakeSampleGroups(summarytable, refgroups = c(), outliers = c())
}
\arguments{
\item{summarytable}{- data frame with columns Sample, Group, etc.}

\item{refgroups}{- character vector. Names of groups that contain "reference" samples.
These groups are relevant for splitting outlier samples into two sub-groups}

\item{outliers}{- names of samples to be considered as outliers. These samples will
be placed in separate groups}
}
\description{
Create a list of vectors describing how samples are related into groups.
}

