% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-makesigs.R
\name{E3MakeSignatureFromScores}
\alias{E3MakeSignatureFromScores}
\title{Create a named list with gene signatures from a matrix with scors.}
\usage{
E3MakeSignatureFromScores(scores, min.score = 0.75)
}
\arguments{
\item{scores}{- a matrix with scores (gene names in rownames and
sample/group names in colnames required)}

\item{min.score}{- minimal score (absolute value). Genes with scores above the
threshold will be considered signature genes.}
}
\description{
Create a named list with gene signatures from a matrix with scors.
}

