% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-sigs.R
\name{E3MeasureGroupConsistency}
\alias{E3MeasureGroupConsistency}
\title{Evaluate various metrics comparing sample and group signatures}
\usage{
E3MeasureGroupConsistency(sampleGroups, sample.sigs, group.sigs)
}
\arguments{
\item{sampleGroups}{- named list of character vectors. Each vector should represent
a set of samples making up a group.}

\item{sample.sigs}{- named list of character vectors. Each vector should represent
a gene set (signature) associated with a sample.}

\item{group.sigs}{- named list of character vectors. Each vector should represent
a gene set (signature) associated with a group.}
}
\description{
Evaluate various metrics comparing sample and group signatures
}

