% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotBasicHeat}
\alias{E3PlotBasicHeat}
\title{Draw a heatmap using Rcssplot.}
\usage{
E3PlotBasicHeat(dat, legend = NULL, ylabels = FALSE, xlabels = TRUE,
  xlab = "", ylab = "", main = "", Rcss = "default", Rcssclass = c())
}
\arguments{
\item{dat}{- matrix of colors.}

\item{legend}{- vector of colors. Legend shading. Warning: the user bares responsibility
to make sure the legend matches with the dat matrix.}

\item{ylabels}{- logical. Toggle display of labels on y axis.}

\item{xlabels}{- logical. Toggle display of labels on x axis.}

\item{xlab}{- character string. Text to display below x axis}

\item{ylab}{- character string. Text to display below y axis.}

\item{main}{- character string. Text to display as title, above heatmap.}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
Draw a heatmap using Rcssplot.
}

