% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotBeforeAfter}
\alias{E3PlotBeforeAfter}
\title{Draw a bar-plot with two bars per label (using Rcssplot)}
\usage{
E3PlotBeforeAfter(before, after, rhs.lab = NULL, col.stim = NULL,
  pixel = 0.1, spacer = 0.04, file = NULL, xticks = seq(0, 250, 50),
  maxlim = NULL, legend.x = c(200, 220), legend.y = 0.8, density = 50,
  legend = c("Preliminary", "Final"), main = "", Rcss = "default",
  Rcssclass = "beforeafter")
}
\arguments{
\item{before}{- numeric vector with names. First set of values to plot}

\item{after}{- numeric vector with names. Second set of values to plot.}

\item{rhs.lab}{- character vector with names. Labels to display next to the bars.}

\item{col.stim}{- vector with colors.}

\item{pixel}{- numeric}

\item{spacer}{- numeric. Determines empty space between boxes.}

\item{file}{- filename}

\item{xticks}{- numeric vector. Determines values to place on x-axis}

\item{maxlim}{- numeric. rescaling factor}

\item{legend.x}{- numeric vector of length two. Determines
the position and size (left and right extemes) of boxes in the legend}

\item{legend.y}{- numberic in range [0,1]. Use to position the legend vertically.}

\item{density}{- numeric. Density of shading lines on one set of boxes}

\item{legend}{- character vector of length two. Text labels describing the
two series of numbers.}

\item{main}{- character string. Title for the plot.}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
Draw a bar-plot with two bars per label (using Rcssplot)
}

