% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotClusteringDiamond}
\alias{E3PlotClusteringDiamond}
\title{Draw a complex diagram containing a tree/diamond clustering and a barplot of signature size
(using Rcssplot)}
\usage{
E3PlotClusteringDiamond(gjh, colmat, gsize, n = 10, spacer = 0.1,
  lab = "All", col.group = NULL, highlight = c(), mai = rep(0.2, 4),
  maxlim = NULL, xticks = seq(0, 300, 100), xticks.dendro = seq(0, 1,
  0.25), lineext = 1.2, multicolor = TRUE, heatcolor = "#282828",
  legend.height = 4, legend.main = "", dendrogram = FALSE,
  dendroheight = 0.8, xlab.right = "Signature size",
  xlab.left = "Jaccard index distance", Rcss = "default",
  Rcssclass = "diamond")
}
\arguments{
\item{gjh}{- hclust object. Used to produce diamond clustering on LHS.}

\item{colmat}{- matrix of colors (or values in range [0,1])}

\item{gsize}{- numeric vector with names. Used to produce barplot on RHS.}

\item{n}{- integer. Sets vertical scale. (Set equal or more than number of elements
in clustering. Fixing this value helps making multiple clusterings with the same
box sizes)}

\item{spacer}{- numeric. Determines empty space between bars on RHS.}

\item{lab}{- character string. Text to write next to clustering on LHS.}

\item{col.group}{- named vector of colors. Associated a color with each element in the
clustering.}

\item{highlight}{- character vector. Names of items to highlight on RHS.}

\item{mai}{- numeric vector of length four. Sets margin.}

\item{maxlim}{- numeric. Determines horizontal scale of barplot.}

\item{xticks}{- numeric vector. Values to label on x-axis in the barplot.}

\item{lineext}{- numeric. determines lenght of line labeling highlighted items.}

\item{multicolor}{- logical. Set TRUE to obtain multicolor heatmap. Set FALSE to plot in grayscale}

\item{heatcolor}{- color for heatmap when multicolor is set FALSE.}

\item{legend.height}{- height of legend (in units as number of bars)}

\item{dendrogram}{- logical, set TRUE to replace the heatmap with a dendrogram}

\item{dendroheight}{- numberic, a relative measure for dendrogram width}

\item{xlab.right}{- x-axis label under bar chart}

\item{xlab.left}{- x-axis label under dendrogram}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
This function produces a figure in two panels. On the left is a diamond clustering.
On the right is barplot showing the signature size for each element in the clustering.
}

