% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotContourScatter}
\alias{E3PlotContourScatter}
\title{Draw a contour plot with select markers}
\usage{
E3PlotContourScatter(x, y, highlight = NULL, col = "#ff0000",
  highlight.label = FALSE, density.transformation = function(x) {     x },
  show.fit = FALSE, xylim = c(0.1, 100), xymajor = 10^seq(-3, 6),
  xyminor = c(), xlab = "", ylab = "", main = "",
  breakrange = range(c(x, y)), numbreaks = 20, Rcss = "default",
  Rcssclass = "contour")
}
\arguments{
\item{x}{numeric vector with names. Coordinates for x axis.}

\item{y}{numeric vector with names. Coordinates for y axis}

\item{col}{color. For highlighting points}

\item{highlight.label}{character vector. Names of points to highlight on the plot}

\item{show.fit}{logical. Set TRUE to display a fitted line through highlighted points}

\item{xylim}{numeric vector of length two. Ranges for axes}

\item{xymajor}{numeric vector. Location of major tick marks}

\item{xyminor}{numeric vector. Location of minor tick marks}

\item{xlab}{character string. Label for x axis}

\item{ylab}{character string. Label for y axis}

\item{main}{character strings. Title of the plot}

\item{breakrange}{numeric vector of length 2. Lower and upper bounds for contour plot}

\item{numbreaks}{integer. Number of partitions for contour matrix}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}

\item{label}{logical. Set TRUE to draw text labels next to highlighted points}
}
\description{
Draw a contour plot with select markers
}

