% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotHist}
\alias{E3PlotHist}
\title{Draw a histogram (using Rcssplot)}
\usage{
E3PlotHist(histobj, xlim = NULL, ylim = NULL, xlab = "", ylab = "",
  main = "", fill = TRUE, xat = NULL, Rcss = "default",
  Rcssclass = "hist")
}
\arguments{
\item{histobj}{- histogram object}

\item{xlim}{- numeric vector of length 2. Determines range to plot on x axis}

\item{ylim}{- numeric vector of length 2. Determines range to plot on y axis}

\item{xlab}{- character string. Text to display below x axis}

\item{ylab}{- character string. Text to display below y axis.}

\item{main}{- character string. Text to display as title, above heatmap.}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
Draw a histogram (using Rcssplot)
}

