% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotLines}
\alias{E3PlotLines}
\title{Draw series data as points with smoothed lines (using Rcssplot)}
\usage{
E3PlotLines(vdat, xcolumn = 1, ycolumns = c(2, 3),
  lineclass = rep("plain", length(ycolumns)), lineloess = rep(TRUE,
  length(ycolumns)), xlim = NULL, ylim = NULL, xlab = "", ylab = "",
  main = "", legend = rep("", length(ycolumns)), legend.at = rep(1,
  length(ycolumns)), Rcss = "default", Rcssclass = "plotline")
}
\arguments{
\item{vdat}{- data.frame with numbers}

\item{xcolumn}{- character/integer. Column containing x values}

\item{ycolumns}{- vector of of columns to show as lines}

\item{lineclass}{- character vector. Classes used to style each lines using Rcssplot.}

\item{lineloess}{- logical vector. Determines whether for each line, the line will
be plotted smoothed (loess) or not.}

\item{ylim}{- numeric vector with two elements. range for y axis}

\item{xlab}{- character string. Text to display below x axis}

\item{ylab}{- character string. Text to display below y axis.}

\item{legend}{- character vector. Text to display in legend}

\item{legend.at}{- numeric vector of length equal to ycolumns. Determines where
to place the legends.}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}

\item{xlim-}{numeric vector with two elements. range for x axis}
}
\description{
Draw series data as points with smoothed lines (using Rcssplot)
}

