% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotScatter}
\alias{E3PlotScatter}
\title{Draw a scatter plot, including line of best fit, correlation measures and labels
(using Rcssplot)}
\usage{
E3PlotScatter(xx, yy, items = NULL, outliers = NULL, col = NULL,
  onecol = NULL, items.highlight = c(), xlim = NULL, ylim = NULL,
  xlimwiden = 0.1, xlab = NULL, ylab = NULL, main = "",
  labelq = c(0.02, 0.98), cex.rescale = NULL, show.correlation = TRUE,
  show.glm = FALSE, show.labels = c(), correlation.threshold = 1e-04,
  Rcss = "default", Rcssclass = c())
}
\arguments{
\item{xx}{- numeric vector with names. x coordinates for points.}

\item{yy}{- numeric vector with names. y coordinates for points}

\item{items}{- character vector. ids of data points to display}

\item{outliers}{- character vector. ids of data points drawn as outliers.}

\item{col}{-}

\item{onecol}{-}

\item{items.highlight}{- character vector. Names of items x,y to highlight.}

\item{xlim}{- numeric vector with two elements. range for x axis}

\item{ylim}{- numeric vector with two elements. range for y axis}

\item{xlimwiden}{- numeric. Determines how much wider the actual x range will be relative
to xlim. Use this to add some padding around the xlim.}

\item{xlab}{- character string. Text to display below x axis}

\item{ylab}{- character string. Text to display below y axis.}

\item{main}{- character string. Text to display as title, above heatmap.}

\item{labelq}{- numeric vector of length two. Quantiles that determine which
items should be highlihted.}

\item{cex.rescale}{- numeric. Rescaling factor for items marked for highlighting.}

\item{show.correlation}{- logical. Toggle display of Spearman correlation info.}

\item{show.glm}{- logical. Toggle display of best fit line.}

\item{show.labels}{- character vector. Names of items to label on the plot.}

\item{correlation.threshold}{- numeric. Correlations with p-value below this threshold
will be highlighted in bold.}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
Draw a scatter plot, including line of best fit, correlation measures and labels
(using Rcssplot)
}

