% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotScatterBubbles}
\alias{E3PlotScatterBubbles}
\title{Draw a scatter plot using bubble of various sizes (using Rcssplot)}
\usage{
E3PlotScatterBubbles(x, y, dsize = NULL, dp = NULL, labels = NULL,
  xlab = "", ylab = "", Rcss = "default", Rcssclass = c())
}
\arguments{
\item{x}{- numeric vector with names. X-coordinates}

\item{y}{- numeric vector with names. Y-coordinates}

\item{dsize}{- numeric vector with names. Determines size of the bubble}

\item{dp}{- numeric vector with names. Determines color shading of the bubble. These
values will be -log10 transformed before being converted into colors
(e.g. use p values associated with each item)}

\item{labels}{- character vector. Names of items which require text labels}

\item{xlab}{- character string. Text to display beside x axis}

\item{ylab}{- character string. Text to display besdie y axis}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
Draw a scatter plot using bubble of various sizes (using Rcssplot)
}

