% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotSegmentationOverview}
\alias{E3PlotSegmentationOverview}
\title{Draws a heatmap representation of a genomic segmentation (using Rcssplot)}
\usage{
E3PlotSegmentationOverview(seglist, chrlens = NULL, categories = list(All =
  names(seglist)), heatcols = c("#00ffff", "#ff00ff"), label.chr = c(),
  seg.rescale = 2, seg.hlab = 0.3, seg.vlab = 1e+06, legend = NULL,
  axes.labels = NULL, Rcss = "default", Rcssclass = "segmentation")
}
\arguments{
\item{seglist}{- list of cbs segmentation data frames}

\item{chrlens}{- named vector with lengths of chromosomes (chromosomes
will appear in this order)}

\item{categories}{- a list splitting the groups into categories (eg. plates)
groups on the vertical axis will be grouped accordingly}

\item{heatcols}{-}

\item{label.chr}{- names of chromosomes to label on horizontal axis}

\item{seg.rescale}{- numeric. Used to modulate color intensity}

\item{seg.hlab}{- numeric. Determines spacing between labels near x axis.}

\item{seg.vlab}{- numeric. Determines spacing between labels near y axis.}

\item{legend}{- numeric. Values to use on the legend box.}

\item{axes.labels}{- character vector of length two. Text to write near
y axis to the left and to the right of the heatmap.}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
Draws a heatmap representation of a genomic segmentation (using Rcssplot)
}

