% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotStimuliDiamond}
\alias{E3PlotStimuliDiamond}
\title{Draw a complex diagram containing a tree/diamond clustering and a barplot of signature size
(using Rcssplot)}
\usage{
E3PlotStimuliDiamond(gjh, colmat, gsize, n = 10, spacer = 0.1,
  lab = "All", col.group = NULL, highlight = c(), mai = rep(0.2, 4),
  maxlim = NULL, xticks = seq(0, 300, 100), lineext = 1.2,
  Rcss = "default", Rcssclass = "diamond")
}
\arguments{
\item{gjh}{- hclust object. Used to produce diamond clustering on LHS.}

\item{colmat}{- matrix of colors (or values in range [0,1])}

\item{gsize}{- numeric vector with names. Used to produce barplot on RHS.}

\item{n}{- integer. Sets vertical scale. (Set equal or more than number of elements
in clustering. Fixing this value helps making multiple clusterings with the same
box sizes)}

\item{spacer}{- numeric. Determines empty space between bars on RHS.}

\item{lab}{- character string. Text to write next to clustering on LHS.}

\item{col.group}{- named vector of colors. Associated a color with each element in the
clustering.}

\item{highlight}{- character vector. Names of items to highlight on RHS.}

\item{mai}{- numeric vector of length four. Sets margin.}

\item{maxlim}{- numeric. Determines horizontal scale of barplot.}

\item{xticks}{- numeric vector. Values to label on x-axis in the barplot.}

\item{lineext}{- numeric. determines lenght of line labeling highlighted items.}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
This function produces a figure in two panels. On the left is a diamond clustering.
On the right is barplot showing the signature size for each element in the clustering.
}

