% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotStimulusLegend}
\alias{E3PlotStimulusLegend}
\title{Draw a legend with boxes and albels (using Rcssplot)}
\usage{
E3PlotStimulusLegend(col.stim, boxwidth = 0.1, file = NULL, spacer = 0.1,
  Rcss = "default", Rcssclass = "diamond")
}
\arguments{
\item{col.stim}{- vector of colors with names.}

\item{boxwidth}{- numeric}

\item{file}{- filename}

\item{spacer}{- numeric}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}
}
\description{
Draw a legend with boxes and albels (using Rcssplot)
}

