% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plot.R
\name{E3PlotVioSet}
\alias{E3PlotVioSet}
\title{Draw a set of violins (using Rcssplot, sm)}
\usage{
E3PlotVioSet(valset, colors = NULL, xlim = c(-1, 1), xat = c(-0.5, 0,
  0.5), highlight = NULL, hidedots = FALSE, showconnect = FALSE,
  xlab = "", ylab = "", main = "", spacer = 0.04, Rcss = "default",
  Rcssclass = c())
}
\arguments{
\item{valset}{- named list of numeric vectors. These vectors are the datapoints
associated with each group and will be used to generate violins. The data points
should contain names themselves (to use highlight option below)}

\item{colors}{- vector of colors. One color for each violin}

\item{xlim}{- numeric vector of length two. Range of x values}

\item{xat}{- numeric vector. Determines where markers will be placed on the horizontal
axis}

\item{highlight}{- vector of item names to highlight}

\item{hidedots}{- logical. Set TRUE to avoid dots on the violins.}

\item{showconnect}{- logical. Set TRUE to connect highlighted items with lines}

\item{xlab}{- character string. Text to display beside x axis}

\item{ylab}{- character string. Text to display besdie y axis}

\item{main}{- character string. Text to display in the title of the plot}

\item{spacer}{- numeric. Determines space between adjacent violins}

\item{Rcss}{- Rcss object. Used to style the heatmap with Rcssplot.}

\item{Rcssclass}{- character vector. Classes to tune Rcssplot formatting.}

\item{hidedots}{- set TRUE to avoid drawing dots on top of the violins}
}
\description{
Draw a set of violins (using Rcssplot, sm)
}

