% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-rescale.R
\name{E3RescaleIntervals}
\alias{E3RescaleIntervals}
\title{Apply interval rescale factors, i.e. change low and high expression estimates.}
\usage{
E3RescaleIntervals(sampledata, samplenames, gene.rescaling)
}
\arguments{
\item{sampledata}{- a list with tables $expression, $expression.low, $expression.high}

\item{samplenames}{- character vector. Names of all samples to rescale}

\item{gene.rescaling}{- numeric vector. Rescales all intervals in both directions.
The function assumes length of this vector is equal to rows of the expression data}
}
\description{
When a gene expression estimate is, e.g.  (2, 4, 6) for (low, middle, high) and the rescale
factor is 1.5, the output expression estimate will be (1, 4, 7). The middle value stays
the same, the distances between the outer values and the middle is rescale by the factor.
}
\details{
Function returns a similar list of table as that in sampledata.
}

