% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-help.R
\name{E3SpearmanRhoFromRanks}
\alias{E3SpearmanRhoFromRanks}
\title{Compute spearman correlations using ranks}
\usage{
E3SpearmanRhoFromRanks(xranks, yranks)
}
\arguments{
\item{xranks}{- numeric vector}

\item{yranks}{- numeric vector}
}
\description{
Compute spearman correlations using ranks
}

