% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-plothelp.R
\name{E3Vioplot}
\alias{E3Vioplot}
\title{Draw violin onto a plot.}
\usage{
E3Vioplot(x, ..., range = 1.5, h = NULL, ylim = NULL, names = NULL,
  horizontal = FALSE, col = "magenta", border = "black", lty = 1,
  lwd = 1, rectCol = "black", colMed = "white", pchMed = 19, at,
  add = FALSE, wex = 1, drawRect = TRUE)
}
\arguments{
\item{x}{- see vioplot from package vioplot}

\item{...}{}

\item{range}{}

\item{h}{}

\item{ylim}{}

\item{names}{}

\item{horizontal}{}

\item{col}{}

\item{border}{}

\item{lty}{}

\item{lwd}{}

\item{rectCol}{}

\item{colMed}{}

\item{pchMed}{}

\item{at}{}

\item{add}{}

\item{wex}{}

\item{drawRect}{}
}
\description{
This is a hacked version of vioplot from package vioplot. This version does not
"box()" every time it is called
}

