% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExpCube-makesigs.R
\name{E3WriteSignatureGMT}
\alias{E3WriteSignatureGMT}
\title{Write signatures into gmt files (text format)}
\usage{
E3WriteSignatureGMT(sigs, expressed.genes = NULL, sig.label = "ExpCube",
  file = "out.gmt")
}
\arguments{
\item{sigs}{- named list with character vectors}

\item{expressed.genes}{- vector with all expressed genes (another signature)}

\item{sig.label}{- character string. Some text used to identify the gene signature within
the gmt file.}

\item{file}{- character string. output filename}
}
\description{
Write signatures into gmt files (text format)
}

